試合登録画面のModel化
